Big Update. Omtænkning af grid & drag and drop funktion. for binding …

+++ src/ShelfMarket.UI/ViewModels/ShelfViewModel.cs
@@ -14,8 +14,8 @@ public class ShelfViewModel : ViewModelBase<IShelfRepository, Shelf>
     private new const string _entityName = "Reol";
 
     #region Form Fields
-    private uint _shelfNumber;
-    public uint ShelfNumber
+    private int _shelfNumber;
+    public int ShelfNumber
     {
         get => _shelfNumber;
         set
@@ -99,7 +99,7 @@ public class ShelfViewModel : ViewModelBase<IShelfRepository, Shelf>
     protected override async Task<Shelf> OnAddFormAsync()
     {
         // Provide default values for locationX and locationY (e.g., 0)
-        var entity = new Shelf((uint)ShelfNumber, ShelfTypeId, 0, 0);
+        var entity = new Shelf(ShelfNumber, ShelfTypeId, 0, 0);
         await Task.CompletedTask;
         return entity;
     }
